requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
curl_cffi==0.10.0
dash==3.0.4
dash-bootstrap-components==2.0.2
Flask==3.0.3
frozendict==2.4.6
frozenlist==1.6.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
multidict==6.4.3
multitasking==0.0.11
narwhals==1.38.0
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.7
plotly==6.0.1
propcache==0.3.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
six==1.17.0
soupsieve==2.7
sutil==0.0.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.0.6
win32_setctime==1.2.0
yarl==1.20.0
yfinance==0.2.58
zipp==3.21.0
